// web/src/api/models/ResponsesUserWidgetData.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mindscape
 * This is the swagger page for the Project mindscape generated with Egg-go. use this to test your database connection
 *
 * The version of the OpenAPI document: 0.0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface ResponsesUserWidgetData
 */
export interface ResponsesUserWidgetData {
	/**
	 *
	 * @type {{ [key: string]: object; }}
	 * @memberof ResponsesUserWidgetData
	 */
	config?: { [key: string]: object };
	/**
	 *
	 * @type {number}
	 * @memberof ResponsesUserWidgetData
	 */
	height?: number;
	/**
	 *
	 * @type {string}
	 * @memberof ResponsesUserWidgetData
	 */
	id?: string;
	/**
	 *
	 * @type {boolean}
	 * @memberof ResponsesUserWidgetData
	 */
	isVisible?: boolean;
	/**
	 *
	 * @type {number}
	 * @memberof ResponsesUserWidgetData
	 */
	positionX?: number;
	/**
	 *
	 * @type {number}
	 * @memberof ResponsesUserWidgetData
	 */
	positionY?: number;
	/**
	 *
	 * @type {string}
	 * @memberof ResponsesUserWidgetData
	 */
	schemaId?: string;
	/**
	 *
	 * @type {string}
	 * @memberof ResponsesUserWidgetData
	 */
	userId?: string;
	/**
	 *
	 * @type {number}
	 * @memberof ResponsesUserWidgetData
	 */
	width?: number;
	/**
	 *
	 * @type {number}
	 * @memberof ResponsesUserWidgetData
	 */
	zIndex?: number;
}

/**
 * Check if a given object implements the ResponsesUserWidgetData interface.
 */
export function instanceOfResponsesUserWidgetData(
	value: object,
): value is ResponsesUserWidgetData {
	return true;
}

export function ResponsesUserWidgetDataFromJSON(
	json: any,
): ResponsesUserWidgetData {
	return ResponsesUserWidgetDataFromJSONTyped(json, false);
}

export function ResponsesUserWidgetDataFromJSONTyped(
	json: any,
	ignoreDiscriminator: boolean,
): ResponsesUserWidgetData {
	if (json == null) {
		return json;
	}
	return {
		config: json['config'] == null ? undefined : json['config'],
		height: json['height'] == null ? undefined : json['height'],
		id: json['id'] == null ? undefined : json['id'],
		isVisible: json['is_visible'] == null ? undefined : json['is_visible'],
		positionX: json['position_x'] == null ? undefined : json['position_x'],
		positionY: json['position_y'] == null ? undefined : json['position_y'],
		schemaId: json['schema_id'] == null ? undefined : json['schema_id'],
		userId: json['user_id'] == null ? undefined : json['user_id'],
		width: json['width'] == null ? undefined : json['width'],
		zIndex: json['z_index'] == null ? undefined : json['z_index'],
	};
}

export function ResponsesUserWidgetDataToJSON(
	json: any,
): ResponsesUserWidgetData {
	return ResponsesUserWidgetDataToJSONTyped(json, false);
}

export function ResponsesUserWidgetDataToJSONTyped(
	value?: ResponsesUserWidgetData | null,
	ignoreDiscriminator: boolean = false,
): any {
	if (value == null) {
		return value;
	}

	return {
		config: value['config'],
		height: value['height'],
		id: value['id'],
		is_visible: value['isVisible'],
		position_x: value['positionX'],
		position_y: value['positionY'],
		schema_id: value['schemaId'],
		user_id: value['userId'],
		width: value['width'],
		z_index: value['zIndex'],
	};
}